Split load into two methods
load_days keeps the functionality of load, and takes
the number of days to load from the past.

It then calls load_date, which now handles all the logic
for checking and storing the data for each date.

load_date can now be called directly from the CLI to
load data for a specific date.